Splitting the code into modules
Refactored the code and split it into several modules.
The core module contains the actual logic to calculate the deltas between mip maps. The other modules contain everything needed for the UI.
Big update and makes it kinda difficult to track changes before and after, but this was overdue, and better now than later.